added school class and attendances concepts

+++ src/main/java/it/pi/registro/registro/dto/response/UserDataResponseDTO.java
@@ -1,14 +1,11 @@
 package it.pi.registro.registro.dto.response;
 
-
 import it.pi.registro.registro.dto.UserSubjectDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDate;
-import java.util.Date;
 import java.util.List;
 
 @Getter
@@ -16,11 +13,10 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserDataResponseDTO {
-
     private String firstName;
     private String lastName;
     private String dateOfBirth;
-    private long daysToBirthday;
+    private int daysToBirthday;
     private int age;
     private String email;
     private String address;
@@ -28,4 +24,5 @@ public class UserDataResponseDTO {
     private String role;
     private List<UserSubjectDTO> votes;
     private String currentSchoolClass;
+
 }
